hailuoai_automator/requirements.txt
pandas>=2.2.3
python-dotenv>=1.1.0
playwright
openpyxl>=3.1.5
#playwright install chromium
loguru>=0.7.3
